+++ spec/publisher_spec.rb
@@ -20,7 +20,7 @@ RSpec.describe PubSubModelSync::Publisher do
 
     it 'supports for custom class name' do
       as_klass = 'CustomClass'
-      inst = described_class.new([:name], klass_name, action, as_klass)
+      inst = described_class.new([:name], klass_name, action, as_klass: as_klass)
       payload = inst.payload(model, action)
       expect(payload.attributes).to match(hash_including(klass: as_klass))
     end
@@ -41,6 +41,20 @@ RSpec.describe PubSubModelSync::Publisher do
       expected_data = { full_name: model.name, email: model.email }
       payload = inst.payload(model, action)
       expect(payload.data).to eq expected_data
+    end
+
+    it 'includes custom headers when provided' do
+      custom_headers = { key: 'custom key' }
+      inst = described_class.new([], klass_name, action)
+      payload = inst.payload(model, action, custom_headers: custom_headers)
+      expect(payload).to have_attributes(headers: include(custom_headers))
+    end
+
+    it 'uses custom_data as the payload data when defined' do
+      custom_data = { id: 100 }
+      inst = described_class.new([], klass_name, action)
+      payload = inst.payload(model, action, custom_data: custom_data)
+      expect(payload).to have_attributes(data: custom_data)
     end
   end
 end