+++ spec/service_kafka_spec.rb
@@ -14,7 +14,10 @@ RSpec.describe PubSubModelSync::ServiceKafka do
   let(:service) { inst.service }
   let(:producer) { inst.send(:producer) }
   let(:config) { PubSubModelSync::Config }
-  before { allow(config).to receive(:kafka_connection).and_return([[8080], { log: nil }]) }
+  before do
+    allow(config).to receive(:kafka_connection).and_return([[8080], { log: nil }])
+    allow(Process).to receive(:exit!)
+  end
 
   describe 'initializer' do
     it 'connects to pub/sub service' do
@@ -31,6 +34,15 @@ RSpec.describe PubSubModelSync::ServiceKafka do
     end
     it 'listens for messages' do
       expect(consumer).to receive(:each_message)
+    end
+
+    it 'subscribes to multiple topics if provided' do
+      names = ['topic 1', 'topic2']
+      allow(inst).to receive(:topic_names).and_return(names)
+      allow(consumer).to receive(:subscribe)
+      names.each do |name|
+        expect(consumer).to receive(:subscribe).with(name)
+      end
     end
   end
 
@@ -61,8 +73,39 @@ RSpec.describe PubSubModelSync::ServiceKafka do
       expect(producer).to receive(:produce).with(match(data_regex), settings)
       inst.publish(payload)
     end
+
     it 'delivers the message' do
-      expect(producer).to receive(:deliver_messages)
+      expect(producer).to receive(:produce).with(payload.to_json, anything)
+      inst.publish(payload)
+    end
+
+    it 'uses defined :ordering_key as the partition_key' do
+      order_key = 'custom_order_key'
+      payload.headers[:ordering_key] = order_key
+      expect(producer).to receive(:produce).with(anything, hash_including(partition_key: order_key))
+      inst.publish(payload)
+    end
+
+    it 'uses defined :topic_name as the topic' do
+      topic_name = 'custom_topic_name'
+      payload.headers[:topic_name] = topic_name
+      expect(producer).to receive(:produce).with(anything, hash_including(topic: topic_name))
+      inst.publish(payload)
+    end
+
+    it 'creates custom topic if not exist' do
+      topic_name = 'custom_topic_name'
+      payload.headers[:topic_name] = topic_name
+      expect(service).to receive(:create_topic).with(topic_name)
+      inst.publish(payload)
+    end
+
+    it 'publishes to all topics when defined' do
+      topic_names = %w[topic1 topic2]
+      payload.headers[:topic_name] = topic_names
+      topic_names.each do |topic_name|
+        expect(service).to receive(:create_topic).with(topic_name)
+      end
       inst.publish(payload)
     end
   end