+++ spec/service_rabbit_spec.rb
@@ -9,8 +9,13 @@ RSpec.describe PubSubModelSync::ServiceRabbit do
   let(:inst) { described_class.new }
   let(:service) { inst.service }
   let(:channel) { service.channel }
+  let(:queue_klass) { PubSubModelSync::MockRabbitService::MockQueue }
+  let(:channel_klass) { PubSubModelSync::MockRabbitService::MockChannel }
 
-  before { allow(inst).to receive(:loop) }
+  before do
+    allow(inst).to receive(:loop)
+    allow(Process).to receive(:exit!)
+  end
 
   describe 'initializer' do
     it 'connects to pub/sub service' do
@@ -37,27 +42,37 @@ RSpec.describe PubSubModelSync::ServiceRabbit do
     it 'listens for messages' do
       expect(channel.queue).to receive(:subscribe)
     end
+
+    it 'connects to multiple topics if provided' do
+      names = ['topic 1', 'topic 2']
+      allow(inst).to receive(:topic_names).and_return(names)
+      names.each do |name|
+        expect_any_instance_of(channel_klass).to receive(:fanout).with(name)
+      end
+    end
   end
 
   describe '.process_message' do
     let(:message_processor) { PubSubModelSync::MessageProcessor }
     before { allow(inst).to receive(:log) }
+
     it 'ignores unknown message' do
       expect(message_processor).not_to receive(:new)
       args = [delivery_info, invalid_meta_info, payload.to_json]
       inst.send(:process_message, *args)
     end
+
     it 'sends payload to message processor' do
       expect(message_processor)
         .to receive(:new).with(be_kind_of(payload.class)).and_call_original
       args = [delivery_info, meta_info, payload.to_json]
       inst.send(:process_message, *args)
     end
+
     it 'prints error message when failed processing' do
       error_msg = 'Invalid params'
       allow(message_processor).to receive(:new).and_raise(error_msg)
       expect(inst).to receive(:log).with(include(error_msg), :error)
-
       args = [delivery_info, meta_info, payload.to_json]
       inst.send(:process_message, *args)
     end
@@ -66,15 +81,39 @@ RSpec.describe PubSubModelSync::ServiceRabbit do
   describe '.publish' do
     it 'deliveries message' do
       expected_args = [payload.to_json, hash_including(:routing_key, :type)]
-      expect(channel.queue).to receive(:publish).with(*expected_args)
+      expect_publish_with(*expected_args)
       inst.publish(payload)
     end
+
     it 'retries 2 times when TimeoutError' do
       error = 'retrying....'
       allow(inst).to receive(:deliver_data).and_raise(Timeout::Error)
       allow(inst).to receive(:log)
       expect(inst).to receive(:log).with(include(error), :error).twice
       inst.publish(payload) rescue nil # rubocop:disable Style/RescueModifier
+    end
+
+    it 'uses custom exchange when defined :topic_name' do
+      topic_name = 'custom_topic_name'
+      payload.headers[:topic_name] = topic_name
+      expect(channel).to receive(:fanout).with(topic_name).and_call_original
+      inst.publish(payload)
+    end
+
+    it 'publishes to all topics when defined' do
+      topic_names = %w[topic1 topic2]
+      payload.headers[:topic_name] = topic_names
+      topic_names.each do |topic_name|
+        expect(channel).to receive(:fanout).with(topic_name).and_call_original
+      end
+      inst.publish(payload)
+    end
+
+    it 'uses :ordering_key as the :routing_key when defined' do
+      order_key = 'custom_order_key'
+      payload.headers[:ordering_key] = order_key
+      expect_publish_with(anything, hash_including(routing_key: order_key))
+      inst.publish(payload)
     end
   end
 
@@ -83,5 +122,11 @@ RSpec.describe PubSubModelSync::ServiceRabbit do
       expect(service).to receive(:close)
       inst.stop
     end
+  end
+
+  private
+
+  def expect_publish_with(*args)
+    expect_any_instance_of(queue_klass).to receive(:publish).with(*args)
   end
 end