+++ spec/spec_helper.rb
@@ -61,6 +61,10 @@ RSpec.configure do |config|
     DatabaseCleaner.clean_with(:truncation)
   end
 
+  config.before(:each, truncate: true) do
+    DatabaseCleaner.strategy = :truncation
+  end
+
   config.around(:each) do |example|
     DatabaseCleaner.cleaning do
       example.run